UI, View Control Inputs: rename 'Control' to 'ViewControl', update descriptions
