+++ Services/Init/classes/class.ilInitialisation.php
@@ -1484,7 +1484,7 @@ class ilInitialisation
                 $c["ui.signal_generator"],
                 $c["ui.factory.input.field"],
                 $c["ui.factory.input.container"],
-                $c["ui.factory.input.control"]
+                $c["ui.factory.input.viewcontrol"]
             );
         };
         $c["ui.factory.table"] = function ($c) {
@@ -1565,18 +1565,6 @@ class ilInitialisation
                 $c["ui.factory.input.field"]
             );
         };
-        $c["ui.factory.input.control"] = function ($c) {
-            return new ILIAS\UI\Implementation\Component\Input\Control\Factory(
-                //NYI
-            );
-        };
-        $c["ui.factory.input.container.viewcontrol"] = function ($c) {
-            return new ILIAS\UI\Implementation\Component\Input\Container\ViewControl\Factory(
-                //NYI
-            );
-        };
-
-
         $c["ui.factory.panel.listing"] = function ($c) {
             return new ILIAS\UI\Implementation\Component\Panel\Listing\Factory();
         };
@@ -1584,42 +1572,42 @@ class ilInitialisation
         $c["ui.renderer"] = function ($c) {
             return new ILIAS\UI\Implementation\DefaultRenderer(
                 $c["ui.component_renderer_loader"]
-                );
+            );
         };
         $c["ui.component_renderer_loader"] = function ($c) {
             return new ILIAS\UI\Implementation\Render\LoaderCachingWrapper(
                 new ILIAS\UI\Implementation\Render\LoaderResourceRegistryWrapper(
                     $c["ui.resource_registry"],
                     new ILIAS\UI\Implementation\Render\FSLoader(
-                    new ILIAS\UI\Implementation\Render\DefaultRendererFactory(
+
+                        new ILIAS\UI\Implementation\Render\DefaultRendererFactory(
                             $c["ui.factory"],
                             $c["ui.template_factory"],
                             $c["lng"],
                             $c["ui.javascript_binding"],
                             $c["refinery"]
-                            ),
-                    new ILIAS\UI\Implementation\Component\Symbol\Glyph\GlyphRendererFactory(
+                        ),
+                        new ILIAS\UI\Implementation\Component\Symbol\Glyph\GlyphRendererFactory(
                             $c["ui.factory"],
                             $c["ui.template_factory"],
                             $c["lng"],
                             $c["ui.javascript_binding"],
                             $c["refinery"]
-                          ),
-                    new ILIAS\UI\Implementation\Component\Input\Field\FieldRendererFactory(
+
+                        ),
+                        new ILIAS\UI\Implementation\Component\Input\Field\FieldRendererFactory(
                             $c["ui.factory"],
                             $c["ui.template_factory"],
                             $c["lng"],
                             $c["ui.javascript_binding"],
                             $c["refinery"]
-                          )
                         )
                     )
-                );
+                )
+            );
         };
         $c["ui.template_factory"] = function ($c) {
-            return new ILIAS\UI\Implementation\Render\ilTemplateWrapperFactory(
-                $c["tpl"]
-                );
+            return new ILIAS\UI\Implementation\Render\ilTemplateWrapperFactory($c["tpl"]);
         };
         $c["ui.resource_registry"] = function ($c) {
             return new ILIAS\UI\Implementation\Render\ilResourceRegistry($c["tpl"]);
