+++ setup/cli.php
@@ -33,7 +33,7 @@ use ILIAS\UI\Component\Input\Field\Factory as FieldFactory;
 use ILIAS\UI\Component\Input\Field\File;
 use ILIAS\UI\Component\Input\Field\Tag;
 use ILIAS\UI\Component\Input\Field\UploadHandler;
-use ILIAS\UI\Component\Input\Field\Control\Factory as ControlFactory;
+use ILIAS\UI\Component\Input\Field\ViewControl\Factory as ViewControlFactory;
 
 $c = build_container_for_setup($executed_in_directory);
 $app = $c["app"];
@@ -222,7 +222,7 @@ function build_container_for_setup(string $executed_in_directory)
             {
                 throw new \LogicException("The CLI-setup does not support the UI-Framework.");
             }
-            public function control(): ControlFactory
+            public function viewControl() : ViewControlFactory
             {
                 throw new \LogicException("The CLI-setup does not support the UI-Framework.");
             }
