+++ src/UI/Component/Input/Container/ViewControl/Factory.php
@@ -5,7 +5,7 @@
 namespace ILIAS\UI\Component\Input\Container\ViewControl;
 
 /**
- * This is what a factory for ViewControl Containers looks like.
+ * This is what a factory for View Control Containers looks like.
  */
 interface Factory
 {
@@ -13,15 +13,15 @@ interface Factory
      * ---
      * description:
      *   purpose: >
-     *      The Standard ViewControl Container is used as the binding element of a collection of Control Inputs
+     *      The Standard View Control Container is used as the binding element of a collection of Control Inputs
      *      for one visualization.
      *   effect: >
-     *      The ViewControl Container is responsible for aligning request-parameters for all contained Controls
+     *      The View Control Container is responsible for aligning request-parameters for all contained View Controls
      *      as well as receiving and distributing values accordingly.
      * ---
      *
-     * @param array<string,\ILIAS\UI\Component\Input\Control> $controls
+     * @param array<string,\ILIAS\UI\Component\Input\ViewControl> $controls
      * @return \ILIAS\UI\Component\Input\Container\ViewControl\Standard
      */
-    public function standard(array $controls): Standard;
+    public function standard(array $controls) : Standard;
 }
