+++ src/UI/Component/Input/ViewControl/Factory.php
@@ -2,7 +2,7 @@
 
 /* Copyright (c) 2020 Nils Haagen <nils.haagen@concepts-and-training.de> Extended GPL, see docs/LICENSE */
 
-namespace ILIAS\UI\Component\Input\Control;
+namespace ILIAS\UI\Component\Input\ViewControl;
 
 /**
  * This describes the factory for (view-)controls.
@@ -13,19 +13,19 @@ interface Factory
      * ---
      * description:
      *   purpose: >
-     *      FieldSelection is used to limit a visualization of data to a choice of aspects,
+     *      Field Selection is used to limit a visualization of data to a choice of aspects,
      *      e.g. in picking specific columns of a table or fields of a diagram.
      *   composition: >
-     *      A FieldSelection uses a MultiSelect Input wrapped in a dropdown.
+     *      A Field Selection uses a Multiselect Input wrapped in a dropdown.
      *      A Standard Button is used to submit the user's choice.
      *   effect: >
      *      When operating the dropdown, the Multiselect is shown.
-     *      The Dropdown is being closed upon submission or by clicking outside of it.
+     *      The dropdown is being closed upon submission or by clicking outside of it.
      * ---
      * @param array<string,string> $options
      * @param string $label
      *
-     * @return \ILIAS\UI\Component\Input\Control\FieldSelection
+     * @return \ILIAS\UI\Component\Input\ViewControl\FieldSelection
      */
     public function fieldSelection(
           array $options,
