+++ src/UI/Component/Input/ViewControl/ViewControl.php
@@ -2,7 +2,7 @@
 
 /* Copyright (c) 2020 Nils Haagen <nils.haagen@concepts-and-training.de> Extended GPL, see docs/LICENSE */
 
-namespace ILIAS\UI\Component\Input\Control;
+namespace ILIAS\UI\Component\Input\ViewControl;
 
 use ILIAS\UI\Component\Component;
 use ILIAS\UI\Component\Input\Field\Input;
@@ -10,9 +10,9 @@ use ILIAS\UI\Component\Signal;
 use ILIAS\UI\Component\Triggerer;
 
 /**
- * This describes the basis of all Control Fields.
+ * This describes the basis of all View Control Inputs.
  */
-interface Control extends Component, Input, Triggerer
+interface ViewControl extends Component, Input, Triggerer
 {
     /**
      * Trigger this signal when the control is being operated.
